agent/requirements.txt
strands-agents>=0.1.0
bedrock-agentcore>=0.1.0